+++ packages/earn-controller/src/types.ts
@@ -5,9 +5,19 @@ export type RefreshStakingEligibilityOptions = {
 export type RefreshPooledStakesOptions = {
   resetCache?: boolean;
   address?: string;
+  chainId?: number;
 };
 
 export type RefreshPooledStakingDataOptions = {
   resetCache?: boolean;
+  address?: string;
+  chainId?: number;
+};
+
+export type RefreshLendingPositionsOptions = {
+  address?: string;
+};
+
+export type RefreshLendingEligibilityOptions = {
   address?: string;
 };